Unwanted position error fixed

+++ src/basic_features/Size.java
@@ -2,20 +2,20 @@ package src.basic_features;
 
 public class Size
 {
-     int x_axis;
-     int y_axis;
+     double x_axis;
+     double y_axis;
 
-    public Size(int x_axis, int y_axis)
+    public Size(double x_axis, double y_axis)
     {
         this.x_axis = x_axis;
         this.y_axis = y_axis;
     }
 
-    public int getX_axis() {
+    public double getX_axis() {
         return x_axis;
     }
 
-    public int getY_axis() {
+    public double getY_axis() {
         return y_axis;
     }
 }
